+++ src/ngraph/distributed/distributed.hpp
@@ -27,8 +27,8 @@ namespace ngraph
         public:
             Distributed();
             ~Distributed();
-            size_t get_size() const;
-            size_t get_rank() const;
+            int get_size() const;
+            int get_rank() const;
         };
     }
 }
