+++ src/ngraph/distributed/distributed_mlsl.hpp
@@ -29,8 +29,8 @@ namespace ngraph
         public:
             DistributedMLSL();
             ~DistributedMLSL();
-            size_t get_size() const;
-            size_t get_rank() const;
+            int get_size() const;
+            int get_rank() const;
         };
     }
 }
